# 🛡️ Fix Critical XSS Vulnerability in FeedParserService (#64)