--- a/.env.example
+++ b/.env.example
@@ -3,3 +3,6 @@ POSTGRES_DB=app
 POSTGRES_USER=app
 POSTGRES_PASSWORD=changeme_secure_password
 POSTGRES_VERSION=16
+
+# HTMLPurifier Cache Directory
+HTMLPURIFIER_CACHE_DIR=/tmp/htmlpurifier