--- a/.gitignore
+++ b/.gitignore
@@ -15,3 +15,8 @@ composer.phar
 
 # GitHub workflow history files
 .github/.history/
+
+###> phpunit/phpunit ###
+/phpunit.xml
+/.phpunit.cache/
+###< phpunit/phpunit ###