--- a/src/Service/FeedParserService.php
+++ b/src/Service/FeedParserService.php
@@ -9,10 +9,13 @@ use Laminas\Feed\Reader\Entry\EntryInterface;
 use Laminas\Feed\Reader\FeedInterface;
 use Symfony\Component\HttpClient\HttpClient;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
+use HTMLPurifier;
+use HTMLPurifier_Config;
 
 class FeedParserService
 {
     private HttpClientInterface $httpClient;
+    private static ?HTMLPurifier $purifier = null;
 
     public function __construct(HttpClientInterface $httpClient = null)
     {
@@ -90,14 +93,46 @@ class FeedParserService
 
     public function normalizeContent(string $content): string
     {
-        // Remove potentially harmful tags and attributes
-        $content = strip_tags($content, '<p><br><strong><em><ul><ol><li><h1><h2><h3><h4><h5><h6><a><img>');
+        // Initialize HTMLPurifier with singleton pattern for performance
+        if (self::$purifier === null) {
+            $config = HTMLPurifier_Config::createDefault();
+            
+            // Whitelist approved elements and attributes per security requirements
+            $config->set('HTML.Allowed', implode(',', [
+                'p', 'br', 'strong', 'em', 'b', 'i', 'u',
+                'a[href|title]', 'img[src|alt|title|width|height]',
+                'ul', 'ol', 'li', 'h1', 'h2', 'h3', 'h4', 'h5', 'h6',
+                'blockquote', 'pre', 'code'
+            ]));
+            
+            // Restrict to safe URL schemes (http/https only)
+            $config->set('URI.AllowedSchemes', ['http' => true, 'https' => true]);
+            
+            // Performance optimization settings
+            $config->set('Core.CollectErrors', false);
+            $config->set('Cache.SerializerPath', '/tmp/htmlpurifier');
+            
+            // Additional security settings
+            $config->set('HTML.Nofollow', true);
+            $config->set('HTML.TargetBlank', true);
+            $config->set('HTML.SafeObject', false);
+            $config->set('HTML.SafeEmbed', false);
+            $config->set('HTML.SafeScripting', false);
+            
+            self::$purifier = new HTMLPurifier($config);
+        }
         
-        // Remove javascript and other dangerous attributes
-        $content = preg_replace('/on\w+="[^"]*"/i', '', $content);
-        $content = preg_replace('/javascript:/i', '', $content);
+        // Sanitize content and handle empty/null input gracefully
+        if (empty($content) || empty(trim($content))) {
+            return '';
+        }
         
-        return trim($content);
+        $sanitized = self::$purifier->purify($content);
+        
+        // Additional security check for malformed content that might slip through
+        $sanitized = preg_replace('/(?:alert|confirm|prompt|eval|console\.log)\s*\(/i', '', $sanitized);
+        
+        return $sanitized;
     }
 
     public function updateFeedFromParsed(Feed $feed, ParsedFeed $parsedFeed): Feed