@dpeuscher 2025-07-18
[AI:Approve-If-OK] Cannot approve PR: API returned 422 error. This may indicate the PR cannot be approved by the current user (possibly the PR author), or there are missing permissions. All other conditions are satisfied - PR is ready, no failing CI checks, comprehensive XSS vulnerability fix with extensive testing.