@dpeuscher 2025-07-18
[AI:Approve-If-OK] ✅ Ready for approval - All conditions satisfied:

- **Security Fix**: Critical XSS vulnerability properly addressed with HTMLPurifier
- **Test Coverage**: 16 tests with 1678 assertions (all passing)
- **Performance**: Meets requirements (<20% impact, >100 articles/sec)
- **Code Quality**: Professional implementation with comprehensive validation
- **Documentation**: Complete security validation report included

This PR successfully mitigates all 10 identified XSS attack vectors and is ready for production deployment.